LICENSE: Add MIT License to code
It was initially made as part of a university assignment. Since it's over, I'd like to open source this code.

I've chosen the MIT License for this repository as while I prefer copyleft licenses for my personal projects, I mostly want this to be for research and archival and am not really bothered if someone steals it.